updated documentation, renamed file
